Use long not int

+++ src/RedisProvider/RedisList.cs
@@ -178,7 +178,7 @@ namespace RedisProvider {
     /// </summary>
     /// <param name="start">Offset index</param>
     /// <param name="stop">Offset index</param>
-    public Task Trim(int start, int stop) {
+    public Task Trim(long start, long stop) {
       return Executor.ListTrimAsync(KeyName, start, stop);
     }
 
@@ -191,7 +191,7 @@ namespace RedisProvider {
     /// <param name="take"></param>
     /// <param name="byKeyNamePattern"></param>
     /// <param name="getKeyNamePattern"></param>
-    public Task<IList<T>> Sort(Order order = Order.Ascending, SortType sortType = SortType.Numeric, int skip = 0, int take = -1, string byKeyNamePattern = null, string[] getKeyNamePattern = null) {
+    public Task<IList<T>> Sort(Order order = Order.Ascending, SortType sortType = SortType.Numeric, long skip = 0, long take = -1, string byKeyNamePattern = null, string[] getKeyNamePattern = null) {
       var getKeys = getKeyNamePattern == null ? null : getKeyNamePattern.Select(s => (RedisValue)s).ToArray();
       return Executor.SortAsync(KeyName, skip, take, order, sortType, byKeyNamePattern, getKeys)
              .ContinueWith<IList<T>>(r => r.Result.Select(v => ToElement<T>(v)).ToList()
@@ -209,7 +209,7 @@ namespace RedisProvider {
     /// <param name="byKeyNamePattern"></param>
     /// <param name="getKeyNamePattern"></param>
     /// <returns></returns>
-    public Task<long> SortAndStore(RedisList<T> destinationList, Order order = Order.Ascending, SortType sortType = SortType.Numeric, int skip = 0, int take = -1, string byKeyNamePattern = null, string[] getKeyNamePattern = null) {
+    public Task<long> SortAndStore(RedisList<T> destinationList, Order order = Order.Ascending, SortType sortType = SortType.Numeric, long skip = 0, long take = -1, string byKeyNamePattern = null, string[] getKeyNamePattern = null) {
       var getKeys = getKeyNamePattern == null ? null : getKeyNamePattern.Select(s => (RedisValue)s).ToArray();
       return Executor.SortAndStoreAsync(destinationList.KeyName, KeyName, skip, take, order, sortType, byKeyNamePattern, getKeys);
     }
