+++ src/RedisProvider/RedisSortedSet.cs
@@ -247,7 +247,7 @@ namespace RedisProvider {
     /// <param name="take"></param>
     /// <param name="byKeyNamePattern"></param>
     /// <param name="getKeyNamePattern"></param>
-    public Task<IList<T>> Sort(Order order = Order.Ascending, SortType sortType = SortType.Numeric, int skip = 0, int take = -1, string byKeyNamePattern = null, string[] getKeyNamePattern = null) {
+    public Task<IList<T>> Sort(Order order = Order.Ascending, SortType sortType = SortType.Numeric, long skip = 0, long take = -1, string byKeyNamePattern = null, string[] getKeyNamePattern = null) {
       var getKeys = getKeyNamePattern == null ? null : getKeyNamePattern.Select(s => (RedisValue)s).ToArray();
       return Executor.SortAsync(KeyName, skip, take, order, sortType, byKeyNamePattern, getKeys)
              .ContinueWith<IList<T>>(r => r.Result.Select(v => ToElement<T>(v)).ToList(),
@@ -265,7 +265,7 @@ namespace RedisProvider {
     /// <param name="byKeyNamePattern"></param>
     /// <param name="getKeyNamePattern"></param>
     /// <returns></returns>
-    public Task<long> SortAndStore(RedisSortedSet<T> destination, Order order = Order.Ascending, SortType sortType = SortType.Numeric, int skip = 0, int take = -1, string byKeyNamePattern = null, string[] getKeyNamePattern = null) {
+    public Task<long> SortAndStore(RedisSortedSet<T> destination, Order order = Order.Ascending, SortType sortType = SortType.Numeric, long skip = 0, long take = -1, string byKeyNamePattern = null, string[] getKeyNamePattern = null) {
       var getKeys = getKeyNamePattern == null ? null : getKeyNamePattern.Select(s => (RedisValue)s).ToArray();
       return Executor.SortAndStoreAsync(destination.KeyName, KeyName, skip, take, order, sortType, byKeyNamePattern, getKeys);
     }
